refactor: update import paths to use new lib structure
