+++ components/index.ts
@@ -5,9 +5,9 @@ import { Footer } from "./src/footer";
 import { CustomMDX } from "./src/mdx";
 import { Navbar } from "./src/nav";
 import { Form } from "./src/form";
-import { OscPost } from "./src/mdxParser/mdxParserTypes";
-import { getPosts } from "./src/mdxParser/mdxParsers";
-import { parseSlug } from "./src/mdxParser/mdxParsers";
+import { OscPost } from "../lib/mdxParser/mdxParserTypes";
+import { getPosts } from "../lib/mdxParser/mdxParsers";
+import { parseSlug } from "../lib/mdxParser/mdxParsers";
 
 export type { OscPost as Post };
 export { CustomMDX, Button, Section, Page, Footer, Navbar, Form };
